+++ apps/meteor/app/utils/lib/slashCommand.ts
@@ -7,8 +7,6 @@ import type {
 	SlashCommandPreviewItem,
 	SlashCommandPreviews,
 } from '@rocket.chat/core-typings';
-import type { ServerMethods } from '@rocket.chat/ui-contexts';
-
 interface ISlashCommandAddParams<T extends string> {
 	command: string;
 	callback?: SlashCommand<T>['callback'];
@@ -133,27 +131,3 @@ declare module '@rocket.chat/ui-contexts' {
 		slashCommand(params: { cmd: string; params: string; msg: IMessage; triggerId: string }): unknown;
 	}
 }
-
-Meteor.methods<ServerMethods>({
-	async slashCommand(command) {
-		const userId = Meteor.userId();
-		if (!userId) {
-			throw new Meteor.Error('error-invalid-user', 'Invalid user', {
-				method: 'slashCommand',
-			});
-		}
-
-		if (!command?.cmd || !slashCommands.commands[command.cmd]) {
-			throw new Meteor.Error('error-invalid-command', 'Invalid Command Provided', {
-				method: 'executeSlashCommandPreview',
-			});
-		}
-		return slashCommands.run({
-			command: command.cmd,
-			params: command.params,
-			message: command.msg,
-			triggerId: command.triggerId,
-			userId,
-		});
-	},
-});
