+++ apps/meteor/app/utils/server/index.ts
@@ -4,7 +4,7 @@ export { getUserPreference } from './lib/getUserPreference';
 export { fileUploadIsValidContentType } from './restrictions';
 export { isDocker } from './functions/isDocker';
 export { getMongoInfo } from './functions/getMongoInfo';
-export { slashCommands } from '../lib/slashCommand';
+export { slashCommands } from './slashCommand';
 export { getUserNotificationPreference } from './getUserNotificationPreference';
 export { getAvatarColor } from '../lib/getAvatarColor';
 export { getURL } from './getURL';
